TD-1136: call linkup api
