+++ apps/dialog/package.json
@@ -60,9 +60,10 @@
     "import-in-the-middle": "3.0.1",
     "ioredis": "5.10.1",
     "isbinaryfile": "6.0.0",
-    "jsdom": "26.1.0",
     "jose": "6.2.2",
+    "jsdom": "26.1.0",
     "katex": "0.16.45",
+    "linkup-sdk": "^2.7.0",
     "mammoth": "1.12.0",
     "markdown-it": "14.1.1",
     "nanoid": "5.1.9",
@@ -92,9 +93,9 @@
   },
   "devDependencies": {
     "@axe-core/playwright": "4.11.1",
+    "@dotenvx/dotenvx": "1.61.0",
     "@eslint/eslintrc": "3.3.5",
     "@playwright/test": "1.59.1",
-    "@dotenvx/dotenvx": "1.61.0",
     "@tailwindcss/postcss": "4.2.2",
     "@telli/typescript-config": "workspace:*",
     "@types/he": "1.2.3",
