+++ apps/dialog/src/app/api/chat/chat-service.ts
@@ -28,10 +28,11 @@ import {
   TOTAL_CHAT_LENGTH_LIMIT,
 } from '@/configuration-text-inputs/const';
 import { ChatMessage, SendMessageResult, createErrorResult } from '@/types/chat';
-import { extractUrls } from './websearch-service';
+import { extractUrls } from '../utils/extract-urls';
 import { UserAndContext } from '@/auth/types';
 import { extractImagesAndUrl } from '../file-operations/preprocess-image';
 import { ingestWebContent } from '../rag/ingestWebContent';
+import { searchWeb } from './websearch-service';
 
 /**
  * Converts frontend messages to ai-core message format
@@ -139,6 +140,9 @@ export async function sendChatMessage({
     urls,
     federalStateId: user.federalState.id,
   });
+
+  // Web search
+  const webSearchResults = await searchWeb(userMessage.content);
 
   // Save user message to DB
   await dbInsertChatContent({
